// Models/picklistcount-model.ts
export class PicklistCountModel {
    BranchId: number = 0;
    CompId: number = 0;
    PicklistDate: string = "";
    TotalPL: number = 0;
    OperatorRejected: number = 0;
    Pending: number = 0;
    Accepted: number = 0;
    Alloted: number = 0;
    Concern: number = 0;
    Completed: number = 0;
    CompletedVerified: number = 0;
}